# OAI/OpenApi/openapi.yaml
openapi: 3.0.0
info:
  title: Tamagotchi Server API
  description: |
    This is the API for a course project at ITMO University. The theme of the course is “Restaurant-application for Android for ordering food and tables”
    [Server on github.](https://github.com/BlackIIIFOX/NodeJsTamagotchiServer)
  contact:
    email: blackiiifox@gmail.com
  version: 0.2.0
servers:
- url: https://virtserver.swaggerhub.com/paul-rozhkin/itmo-tamagotchi/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://petstore.swagger.io/v2
- url: http://316825-blackiiifox.tmweb.ru:3000/api/
...
paths:
...
  /files/:
    post:
      tags:
      - files
      summary: Upload new file to server.
      operationId: uploadFile
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        "201":
          description: File uploaded.
          content:
            application-json:
              schema:
                $ref: '#/components/schemas/fileInfo'
        "400":
          description: The file is not in the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
      security:
      - bearerAuth: []
  /files/{id}:
    get:
      tags:
      - files
      summary: Download file by {id}.
      operationId: downloadFile
      parameters:
      - name: id
        in: path
        description: File id.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: File
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        "404":
          description: File not found by id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    fileInfo:
      type: object
      properties:
        id:
          type: integer
          description: file id
          example: 2123
        name:
          type: string
          description: file id
          example: newFile.jpg
        type:
          type: string
          description: file id
          example: image/jpg
...
    errorModel:
      type: object
      properties:
        message:
          type: string
          description: Error description.
...
    body:
      type: object
      properties:
        file:
          type: array
          items:
            type: string
            format: binary
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT